Fix up makefile and refactor mem test
Now mem test doesn't wait in lockstep on each command for the result to come back, greatly increasing potential throughput. On the real hardware this drops from about 48 cycles/read to about 12 cycles/read, so we should def be using this pattern and make sure the sim implements it properly.