Fix netlist explosion(!!!)
Big issue here was when using the regs and regs_next two-process
pattern, the quartus compiler didn't seem to be able to understand that
we meant to have a register file and only replace single elements at a
time. This new approach separates out some write and data signals and
adjusts the logic in the sequential block to make it more explicit, and
now the netlist looks a _lot_ better, at the cost of a _tiny_ amount of
complexity in the new signals, but we also get prettier reg write code
throughout the writeback state.

This patch also applies the same logic to the read buffer, which doesn't
seem to have much affect in that case.