Remove some redundant error state transitions in CPU
Simplifies net a bit, but at the cost of a few more compiler warnings. I think I'll end up converting a lot of this logic to if's to get around this in all compilers, but I'll wait until I refactor/simplify loads/stores in general before I do that, as a lot of these cases will go away then anyways.