Make sure r0 is 0
This is done by making two changes:
1. Make register file regs array only go down to index 1
2. Add addr compares to the read_data paths, so that if the addr is zero, the value should be zero as well, else read from the array.

Together this appears to add a small amount of logic driving the WE line for the inferred RAM, as well as a 32-bit mux on each read port. This isn't ideal, but is still pretty self-contained, and I didn't feel comfortable letting the RAM contents be unitialized or relying on the output ports of the inferred RAM to behave properly when the read addr was zero. It's a bit of a shame that the inferred RAM doesn't have a proper clear to zero line that we could use on reset, but this alternative isn't terrible.