Separate store unit from cpu
This greatly simplifies the logic used for unaligned stores, and is generally pretty nice within that unit. The unfortunate thing here is that the design grows a bit in order to accomodate this separate unit with the existing cpu code, which also needs to be changed a bit more fundamentally (especially wrt bus arbitration), but I don't really think I can fix that without hitting the instruction fetch and memory load units first and changing it all as a whole. But I think it's a good idea to do it a bit piecewise in order to see different tradeoffs etc with how different pieces _can_ fit together.