Separate cpu into cpu and core modules
Big thing here is that while separating out the store unit from the rest of the cpu, it started to become quite obvious that the mem bus arbitration didn't belong in the core cpu logic at all, even though it was necessary to have somewhere in order to ensure correct/non-conflicting bus usage. It also seemed wrong that I had to name things around certain port names that were really just there to instantiate the store unit, even though the core itself only used a few of these signals. So what _was_ the cpu is now the _core_, and the _cpu_ is just a high-level module that instantiates the core and store unit and does some trivial bus arbitration. It's most likely that when I split out the load unit and probably the instruction fetch unit later as well that a separate arbiter unit will make sense as well, but this already feels like a step in the right direction with how things fit together and should hopefully get simpler.