Separate out load unit from core
Design got a bit bigger again, probably due to the extra muxes and whatnot related to bus scheduling, but yeah. Code is a bit easier to understand overall, and the loads and stores individually are much better than they were before. I can still separate the instruction fetch unit and possibly reduce some more logic here, but I might move on for a bit as I'm really keen on getting this talking to DDR mem.