Add UART transmitter
This was rather straightforward actually, as it should be :) seems to work quite well, although my current ROM test seems to leave out the first char in all strings sent except the first one, and that's pretty strange. I don't think it's anything to do with the UART itself, probably something in the higher-level code, but we'll have to see.

Also note that this hasn't been implemented in the simulator yet; that
may help debug actually :P