Start fleshing out other pipeline stages
First instruction works (JAL) but not in a way I'm happy with yet. One nice thing is I realized we don't need separate EX/MEM stages, so those have been merged. Also, since most stages will just use inputs determined in the decode stage, we don't need to actually make additional registers between the other stages; we can do that when (if\!) we move to proper pipelining. Looks like most of the bulk of the work is going to be done in the execute_mem stage (I guess that's to be expected).