More progress
Mostly moving all of the complicated execution/decoding into the EX/MEM stage, which seems to be going OK. Branch addresses look like they'll work, starting to understand how the loads/stores will be issued, and the writeback stage should end up being mostly just pass-through/combinational stuff to commit results from the previous stage.

Interestingly enough the implementation appears to be quite a bit larger than the previous implementation already (before being complete which is somewhat surprising). It won't be a practical issue (we're talking an additional few percent of the overall area of the FPGA here), but I would have liked it to be smaller in the end. It might still be if I can come up with some more simplifications. In any case, even though it's larger, the RTL viewer is showing much more straightforward stuff, and I think that's a really good sign.