Replace control with python HDL
Interestingly quartus no longer recognized a state machine, and even though it now uses one-hot encoding, we have more logic elements (this may be due to the fact that state is now larger; might want to play with not using one-hot as well).

Seems to be a pattern emerging for logic for such a state machine that, if it seems to repeat, should lend itself to a nice abstraction yay.