Basic immediate loads+jal are working
Plus most other instructions in theory, but they're not tested yet. Best part is that the RTL viewer in quartus seems quite happy with the input; the ex/mem stage ends up being a bunch of simple binary op's and muxes just like it should be, so I'm quite happy about that.