Mem map doc overhaul
The main difference is removing mirror definitions, as while these are likely technically correct, I don't want them to be relied on. I'd like reads/writes to everything other than the explicitly specified registers to be undefined so we have more freedom with the actual implementation (and specifically, the freedom to change which mirrors are where if it makes the implementation of the actually relevant parts of the memory map simpler).