Hook up HDML SCL/SDA lines and display interface, start bit-banging
This code ends up being pretty straightforward, which is nice, but I'm having issues in test again that look like timing errors. It's timing errors in the same places (internally in the memory controller) which is, needless to say, extremely frustrating. It does indeed look like those errors have gotten worse so I'm not surprised that that actually may be the issue but yeah.