Check new DDR3 controller in signaltap
Reads still appear to have the same bandwidth as before; I hope stores are better. What I should probably do is work towards actually getting a bunch of stores in bursts working, and get a tap of that, to see what we're really working with. Perhaps implementing a hardware clear is a great way to do that.