Remove DDR3 controller altogether, use internal BRAM instead
This is mostly to get around all of the annoying controller timing issues I've had, and is only a temporary measure until I get a new board. The biggest downside now is that we only have 128KB of RAM instead of 128MB (we could have a _bit_ more but I want to still be able to use signaltap when we need it), so I dropped the framebuffer/output resolution to 160x120. Additionally, to make things easier (and since we no longer are stuck at 150mhz just to interface with the DDR controller), I've clocked down the whole design to 100mhz, and retimed things. This means we have a _lot_ of slack, and even better, I added some design constraints so that things stay that way.

Besides the lower specs we're at full feature parity with what we had before this, at least for the rtl/rom. The simulator needs to be updated still, but that's more or less trivial.