Set up uart test project
This is important for several reasons; first of all, it's the first successful test project on the new board, yay! It also sets the standard for how I want to work with vivado files in the repo; I think I'm pretty happy with this.

Two things I'm not totally comfortable with yet:
 - The design constraints (particularly the output delay for the tx pin) needs to be reviewed/revisited when I understand this better. I don't want to totally unconstrain the path but I'm not entirely sure what I should use for this particular pin, and I don't fully understand the min/max args yet.
 - I had to add 'wire' to a few of the port declarations in existing rtl code, as the vivado compiler seems to be a bit more pedantic. It should be reviewed whether or not this is the right fix for the code (vs eg. logic something or other).