Fix up design constraints and code style for uart test project
After some deliberation, I decided that for uart tx, we actually don't need constraints, since it's an asynchronous signal anyways, so we can make that a false path with dummy delays like the input reset signal, and improve the comment to clarify this. As for default_nettype with the vivado compiler, I want to use 'wire logic' instead of just 'wire' or 'logic' to be as explicit as possible here, so I've fixed up the relevant (used) source files, and will fix up additional source files as we add them to their respective test/top-level project(s).