Improve reset synchronizer implementation
This is more basic SV, but good practice for me. The first change is to accept a number of stages as a parameter, and clean up the code while I'm at it. The second major difference is to mark the sync chain regs with an inline attribute that helps the tool understand their purpose and place them close together so as to give the signal between the regs in the chain as much slack as possible, which allows them to better perform their jobs.