Swap out hello message with lfsr for uart tx test
This is something I had done before, and so far it looks really good. The next step will be to make the rust program loop back the data, so we can also perform the check on the device, which allows us to test the UART receiver as well as the transmitter.

One thing to note that I also changed was to update the LFSR state after checking the next value instead of before - this makes the digital logic impl slightly simpler, and is nearly identical but with just a single value phase shift at the beginning of the sequence, which isn't important at all.