Complete uart lfsr loopback setup
Now instead of a single write datapath, we also add a parallel read/compare datapath with its own lfsr to compare against. I've also added an LED for error output, and I've configured the test to stop sending data if an error is detected, which should help us identify how many bytes were successfully round-tripped if an error actually occurs. Now I can run this as long as I can, and we have the output of the lfsr checked both on the PC and on the input path of the device.

One crucial thing that I also did was to change kaze to be explicit about declaring the net type and data type of the inputs/outputs in generated code, which was required to get the uart_receiver module to be compatible with the vivado compiler with default_nettype none declared. I suspect this will be the only change we'll have to make with kaze, but we'll see if there are more.