Set up DDR3 controller test
This test performs a big bulk write of all 256MBs of available DDR3 as fast as possible while counting cycles, transmits this cycle count over UART, and then does the same for reads. This allows us to get exact cycle counts for each phase, and it looks like we're able to achieve very close to a read/write every single cycle, as expected. It's likely less due to periodic refresh cycles and/or controller overhead; it doesn't really matter, this behaves as well as I could have hoped, which validates the planned DMA approach, so I'm very pleased.

I also decided that the test programs that run on the PC to monitor/communicate with the test via UART should be placed in the test dir instead of closer to the root dir, so I moved the lfsr-test project into the uart test dir as well.