Simplify uart_clock_divider
This is a bit more trustworthy, although the UART is still slower than I think it could be. I actually wonder if it might have to do with the fact that we don't have any timing constraints for tx/rx (we currently mark them as false paths) - this should be relatively easy to test, but I seem to have misplaced the power adapter for my board, so I'll just leave a note here for now and focus on some other tasks in the meantime.