Start porting cpu -> marv
I'm now introducing the name MARV, or Mostly Adequate RISC-V. This is the first big module to be reimplemented in the new kaze rust impl. It's not complete yet; most if it's going fairly smoothly, although I already ran into one use of lt_signed that's not implemented in kaze yet, unfortunately. However, the top-level code is looking much nicer than the equivalent .sv file, which I'm quite happy with (even though to get around similar missing signed binops, we're going to want to use an external alu at first).

My short-term goal is to get code running in a simulator that consists of a generated marv sim and rust bootstrapping. Once that works, we have a serious proof of concept on our hands (even though we don't generate rtl atm still, hehe).