MARV appears to work flawlessly!
Some small changes in the top-level bootstrapping and some small simplifications I found while doing other things, and I also uncommented more ROM code for extra testing. One thing is that this actually doesn't work with the currently-released kaze, as it turns out there's a bug in kaze_sugar which the MARV code relies on quite heavily. I had to do some really nasty things to get the current syntax to work, but it was worth it to test if the MARV logic was correct, and in fact, it totally is, which is just awesome. I'll need to spend some time figuring out what to do about the kaze_sugar impl/syntax, but this was expected anyways, and is why it's marked as 'unstable' in the API docs in the first place.