Use CamelCase for module names, and snake_case for instance names
The only real difference here is that now there's no ambiguity when an instance is called the snake_case version of its Module name, where before, I would suffix an instance with an index (usually 0). This also means there's no need for allowing snake_case module names in kaze, which I think is a pattern I'll lean into for that lib.