Remove offset from initial stack pointer
Looking at disassembled ROM code, we can see that like x86 and many other platforms, the stack pointer is decremented before use, which means we're wasting a word by offsetting its value in the linker script. Instead, we should set its initial value to the location just past the end of memory.

We can verify that this is behaving correctly by making mem one word smaller in marv-test, and see that it crashes with an OOB memory access when using the stack.