Experiment with drive_next pattern, simplify mem stage
This stage is still mostly just pass-through, and these regs should probably be lifted out into marv, but for now we can still work on simplifying this kind of code, which I think will be common. For now I'll continue to have a reg_next impl here, and we can replace it when I make it a proper feature in kaze.