Refactor register file for synchronous reads
I've decided that this is how reads will work in kaze, so let's test that here first. Now the decode stage is perhaps a bit more aptly named since the rs1/rs2 values are decoded as the data is returned in order to feed the register file read ports in time for the data to be ready in the reg wait stage. I may refactor a bit to put this small amount of decoding actually into the Decode module.