Add peek buffer
The main idea here is to provide a unit which will optimistically read from a FIFO (or some other thing that returns data on the cycle following a successful read handshake) and provide that data immediately (if possible) or hold it in a buffer so that it can be retrieved later. Since it always outputs the data it's holding, this means that we can see what the data is (hence "peek") before we decide to actually take the data (which would cause the buffer to read in more data from the FIFO, if any).

The main use case here is as part of the Buster unit, where we need to see which replica FIFO we're going to read data from next before we can actually read the replica data and primary return FIFOs. With this unit, we can wait until the peek buffer can tell us which replica buffer to pull from, and as soon as we have that data and the other FIFOs are ready, we can simultaneously issue a read to each one (including the peek buffer) and all of the relevant return data will be picked up in the next cycle (as well as new data into the peek buffer if any, which should be able to maintain peak throughput).

Important to note is that it assumes that its ingress ports are always connected to something that will return data on the cycle following a successful read handshake; anything that violates that assumption will not work correctly with this unit (which is totally OK).

I'm not 100% sold on the name here, but I felt it wasn't quite a skid buffer, even though that's probably the closest thing (but I'm not 100% sure of that either since looking up what a skid buffer actually is seems to yield wildly different results, so I'd rather come up with a new name instead of reusing an existing one when I don't think it's the same concept as what I've previously seen). In any case, this is *a* name which is a starting point.