Make replica_select_bit_width a param for Buster
I was originally thinking to derive this based on the number of replicas used, but that would mean that adding/removing replicas would change the memory map. In theory we could also generate headers for the map based on the Buster hierarchy, but it doesn't sound that appealing to make, so I'd rather just conservatively use a few extra matching bits (which I can segment in logical, easy-to-remember ways like 4 bits per hierarchy level) and specify which explicitly.