Parameterize number of replicas in buster
Unfinished, but I wanted to push what I have now so I can relax about it. I'm also going to want to refactor all of this including the primary stuff when it's all done and tested more, as it's getting a bit messy, especially with all the conditional stuff. But I'm still thrilled that it's possible to do this at all, and I think this will be a really nice module to work with.