Add Buster2x1 tests, fix build for num_replicas = 1 case
Super unhappy with how the code looks atm - but also super stoked that it works! I'm glad I allowed myself to make a bit of a mess while getting this to work, but I really need to take some time to massage it into something more manageable, especially since this is a very critical component that will see a lot of (re)use in the design.