Hook up first top-level xenowing test
This harness hooks up the xenowing module which consists of a MARV instance as well as some interconnect, and exposes bus signals for all of the peripherals (for now; these will eventually also be implemented in kaze). Since this interconnect is now built on buster instances, the memory map has been changed to reflect the new scheme this imposes, and the riscv-compliance suite now passes with an updated linker script (required as the PC reset value has been changed). The ROM code (which I think I'll rename to BIOS soon) has also been updated, and works with the top-level sim harness, which is awesome!

I did have to fix one bug in the ROM code for put_u64 which I have a hard time explaining, but the new code is clearer anyways, so I'll ignore that for now.