Port UART transmitter to kaze
Also clean up lots of stuff, including removing all of the old rtl for UART transmission. This probably breaks the ddr3 test, but we'll do a pass on that to update it later anyways.

Now there's nothing holding us back from describing the whole UART test module in kaze except for filling in features for kaze's .sv codegen, so we'll dogfood that next.