Hook up proper UART interface/tx in xenowing module
Also add a new top-level module to the xenowing sim test called Top, which instantiates a Xenowing module, and hooks its UART tx to an instantiated UartRx receiver module so we can reuse that decode logic, and then exposes the data and data_ready ports so we only have to monitor those in the sim setup. In the hardware setup, we'd just route this tx to an output pin directly, but this ensures that we expose the right interface from the xenowing and also that it always interacts with a transmitter that has a proper output delay, which is important for software correctness.