First successful host file IO test!!!!
This is a big one! So far only tested in sim, but since all of the host code is written against an interface that just reads/writes bytes from/to the xenowing, this should work over actual UART as well (in fact the simulated system can't tell the difference; it encodes/decodes UART for all of this data anyways).

I did, however, observe that if I send a u32 with the MSB set, the xenowing would hang before it printed the file size. I have no idea why this happens, but it was consistent, which is odd.