Rename interconnect -> interconnect0
As it turns out, 'interconnect' is a SystemVerilog keyword, so we can't use that here. I'm a bit on the fence about trying to enforce this kind of thing at the library-level since it's hard to make that kind of thing watertight, so I think I'll just change it here in user code instead.