Merge xenowing sim/xw-blaster, split ROM
Now ROM is split into boot ROM and program RAM, with different memory areas for each, and we can use the xenowing blaster to connect to either a simulated device or a real one over serial.

Additionally, this refactoring "magically" fixed some load time issues for the real device, which I guess was due to software overhead.

Note that there's a lot of duplication here and I'm not 100% happy with the code yet, but everything works and is finally updated and working on real hw, so this is a great checkpoint!