Separate pixel pipeline from the rest of the color_thrust module; quick blending test
This separation cleans things up by separating inputs to the pipe and other signals coming out of the pipe. It's slightly more complex due to more plumbing/pipeline stage regs/extra active signals (since the pipeline may still be active after the input stage has finished generating inputs). I've also implemented a basic blending test which verifies that we can multiplex the color buffer read port and use it to mix with the input color of the pipe (which means we don't have to double-buffer like I originally thought); this also verifies that multiple pipeline stages works as expected, and in particular that we correctly update the pipeline's active state, which goes high when the start signal is asserted, and goes low once a `last` signal has been propagated through all of the pipe stages (this is sent from the input generator when it's active and generating inputs for the last pixel in the tile). This is also a good test to see how this kind of blending might look in rtl. Overall, very happy about this!