Add WordMem, use it to implement program ram and ddr3 mem
A common pattern that comes up a lot is a memory that's organized into words made up of several elements - for example, 128 bit words as 16x8bit elements, where we want the write port to support an individual write enable signal for each element in addition to the port's global enable signal. I've implemented this so far by making a separate kaze Mem object for each element and splitting/combining all of the word/element signals manually, and it looks like I'll need this in even more places, so I think it makes sense to add an abstraction for it. Additionally, I chose user-defined element size instead of byte elements (which are the most common) since I can see this being useful for other sizes as well, for example, per-pixel elements for texture cache mem, where individual bytes won't ever be written.