Implement bilinear filtering for pixel pipe
This is fairly involved since we need to be able to perform 4 texel reads in a single cycle in order to filter properly. To achieve this, I split the texture mem storage into 4 separate memories, and organize the texture as a series of 2x2 blocks. Each pixel in each block is stored in a different memory and interleaving is expected to be done as part of texture preparation. This way, we can still compactly store textures that are smaller than the maximum size, so long as all of the relevant 2x2 blocks are packed properly (so we'll still be able to fill the texture cache with a single DMA operation). In order to read from the correct locations of each of the 4 memories we have to potentially adjust the read addresses as well as the blend weights depending on the lsb's of the texel addresses, but the logic for this is actually relatively straightforward and the rest of the pipeline is basically untouched.

One surprising thing was that we were able to meet timing just by swapping blend weights in the stage before blending and separating the s and t blending parts into two stages; I suspected we'd at least need to pipeline the multiplies, but I guess since the multiplies are so small (8x5 and we drop 5 result bits) it's implemented as normal logic instead. This also means there's less DSP pressure and I may not even need to refactor blending like I originally thought (as a + (b - a) * w instead of a * (1 - w) + b * w which is conceptually simpler and only uses unsigned multiplies which is all that kaze supports currently).