Add initial strugl test
I guess strugl is the temp name for the graphics API for the system. Currently though, it's a mixture of the current xw-blaster test and the previous gloat setup code so that I can write basic 3D-oriented tests and simulate only a ColorThrust instance rather than the whole xw system, which is much faster. I've also massaged gloat's model to more closely match the current rtl impl. It's not a perfect match (mostly due to me being lazy about texture filtering) but it's enough to get a good comparison and was integral in identifying and fixing the inverse reciprocal pipeline bug I just fixed.