Expose 128-bit texture write path in Device abstraction
This moves some complexity around a bit but in total seems a bit simpler actually. The main benefit here is that we get to reduce the amount of cycles we need to simulate during texture data transfer, since we move more data per cycle. This is the same optimization that the DMA approach will accomplish eventually in the system, but in this sim test setup, it's also useful to save host CPU cycles and allow the test to run smoother.

While optimizing the texture write path doesn't really gain any noticeable performance improvements, it's the easiest part to do so I thought I'd start there. The other paths should be similar.