Only move depth buffer tile data in/out depending on depth settings
We only need to move data into tile mem if we're going to use it, either by testing against it or writing to it. Further, we on'y need to move it out of tile mem if we've written to it, which only occurs when write mask is enabled. Even if we move depth buffer data into the tile for depth testing, we may not have updated it, and wouldn't need to move it back out again.