Initial test tile triangle binning impl
This duplicates a TON of data and feels pretty inefficient for now, but already in this simple test, significantly reduces the amount of time spent transferring data, leading to a 20% reduction in estimated frame time. In reality we'll gain a lot less than that due to the overhead of writing all of the bin data and reading it back again, but we can cover some of that latency by updating rasterization regs while the rasterizer is working, which we should be able to try now actually.