Mirror dx/dy regs for interpolants to reg update/rasterization overlap
This is a more or less trivial change in the rtl that allows us to change all of the rasterization parameters while rasterization is actually happening. This was already the case for min values which were only copied to the internal counters when the start signal was asserted, but we had to add extra registers for dx/dy.

In the very simple two-cube sim test setup, this already saves a couple thousand extra cycles, and those gains should be even more significant with real workloads.