Initial read cache impl + ColorThrust doc
The read cache will eventually be used to implement a proper texture cache in ColorThrust, in addition to an I$ for the CPU (probably). It's not complete (still needs a way to trigger invalidation under normal operation, and probably more testing) but it's finally coming together after months of hard work, kaze features, and ofc some procrastination as well. :)