Remove reg wait pipeline stage in marv
Since the register file uses synchronous memories with enable signals for the read ports, we can just use the registered outputs there and don't need to add an additional cycle of latency to compensate for poor timing with asynchronous read memory like we had before. With this change, we've got the same classic 5 stages that the RISC pipeline does, which should simplify adapting pipelining logic later on (though there's one crucial difference here; writes to the register file won't be visible to reads to the same registers on the same cycle, so we may have to add some additional stall logic to cope with potential r/w hazards which the original pipeline doesn't have).