Update color thrust pipe doc
This includes a separated depth pipe so we can do that test right away, and potentially avoid shipping pixels into the texture pipes. Additionally, the input generator can choose to discard pixels based on the edge test, which we can do before we even push to the depth pipe. So we should be able to drop a lot of pixels, potentially even behind a tex cache stall, which is pretty rad.