Reject pixels with edge test before they enter the pixel pipe
This is more or less trivial after the last commit. Now we just determine a reject signal and gate the valid signal going into the pipe if the pixel is rejected, as well as modifying the count that's added to the finished pixel accumulator. All other control logic is the same.

Another benefit of this is that we can drop the edge test signal from _all_ pipe stages, because all of the mem reads/writes are already gated by the valid signal, which itself has already been gated by the edge test. Nice.