Initial texture cache impl
A few loose ends that will have to be co-developed further with the software/test setups, but the major parts are there, and it works!

Note that I haven't tested on FPGA yet (I'm on holiday and don't have hw with me) but I suspect it will work just fine; perhaps we'll need to add a stage or two to the split pipelines if timing is hard to meet.