Make Context generic over Device
This is in contrast to taking a mutable reference. This will allow this abstraction to be zero-cost for cases where the Device type is known at compile type, and by adding a Device trait impl for mutable references to Device impls (mimicking the pattern used by std::io::Write) we can continue to use dynamic dispatch as well.