Get xw-blaster "working" again
The main thing here really is that the color thrust replica bus is now connected in the xenowing module, and probably works, but it's kinda hard to tell beyond "it doesn't lock up" because it just produces black triangles in the full sim setup (and has for a long time, and I won't fix it because I want to merge these two test setups anyways).

Additionally, the strugl setup has been updated to wrap the ColorThrust module in a Top module which includes memory internally, and a buster crossbar to arbitrate/manage access to that memory between an externally-visible bus and the ColorThrust's replica bus.