Fix xw-blaster build and tex cache addressing
I somehow broke these when refactoring, which turned out to be a PITA to debug. However, I believe this was a net positive, as I had to rigorously go through and document various address widths through the signal path to get it right, resulting in helpful notes and fewer magic constants. Yay!