Beginnings of a new Rust boot rom
I'm not super thrilled with how all the pieces fit together. For example, the boot-rom crate has to be excluded from the top-level cargo workspace because I want to have specific target options for it (to reduce its size), and I can't get any projects to actually depend on it (likely due to it only building for another target), so it needs to be rebuilt manually just like the previous boot_rom. However, it's already a bit nicer, since the top-level Makefile actually builds it now!

Still some stuff to do, like getting an allocator set up, but so far it's working well even though it's imperfect.