Fix buster bus_ready propagation bug
When the number of bits to select which replica to target is 0 (which is sensible when there's only one replica), the previous logic would always propagate high for bus_ready, which would cause primaries to see valid handshakes when they weren't supposed to. This was causing stalls as the CPU would issue reads to memory, see a "valid" handshake, and wait for the result; the memory, on the other hand, didn't see a valid handshake, and would never return data. Welcome to stall city.