strugl: Adjust bin cycle estimate
So this was a big oversight: I've been assuming each _byte_ would cost a cycle to read/write for bin data, but we can of course use much higher bandwidth than that. Now, we at least assume we can process 32 bits/cycle, which I think is realistic, assuming we have some kind of FSM that shifts 32-bit quantities in and distributes them to various rasterization registers (even though that's only a quarter of our available memory bandwidth).