rtl, xw-blaster: Expose XenowingInner module, add sim-inner option
The main thing here is to expose an 'inner' Xenowing module with its uart interface exposed directly, rather than hooked up to an actual encoder/decoder. This provides a dramatic increase in sim test turnaround times, where a lot of time is typically wasted just waiting for data to be transferred over uart, when most of the time, this part of the system isn't particularly interesting to test.