Implement MIG UI DDR3 sim, use in sim and sim-inner xw-blaster modes
Lots of code copypasta here, but I'm not sure how to refactor this in a satisfying way (eg. just stuffing all the ddr3 stuff into its own struct with a messy interface doesn't really cut it), and frankly it's just not that much logic that it makes much of a difference to me, so let's just roll with it and move on.