Rename color-thrust-meta -> rtl-meta
It's a bit tiring and not actually that helpful to have to make 2-3 separate projects every time I want to incorporate a new "silo" of hw, esp. when I'm the only one working on this stuff and none of it is being used externally.

Also, suddenly we have a natural place for some duplicated stuff (eg. the number of system address bits), which I should clean up soon.