Start on model BitPusher, use in strugl to move tile data
This ofc breaks everything but the model until I've implemented the RTL. There's also a "write a const u32 a bunch of times to memory" mode that I haven't covered. However, I think the most interesting and fun thing to do next is reach feature parity with RTL with this first, and then implement the const mode in a second pass.