Start shovel-test
The idea here is to get shovel up and running in a realistic enough sim environment to hack on the rtl without fpga turnaround times. So far the video test pattern generator appears to work exactly as expected, so the next step is to simulate a proper system clock frequency and space out the video domain control pulses properly (enough).