shovel: Rebuild gateware after core template upgrade
Interestingly, we no longer meet timing, but we only fail on a few paths. This is likely either just noise or due to the bridge command logic taking up a bit of space; I think the former is more likely. In either case, the easiest thing to do now is just reduce clock rate slightly again.