rtl/sw: Refactor BootRom into Bootloader
The main idea is that the bootloader should execute in its own little self-contained memory corner so we can iterate with it to help with SDRAM bringup (and perhaps initialization later, if we want/need to do that from software). The pocket's asset (re)loading system can be used to (pre)load the data before the system comes out of reset, and we can just hook up a mem with initial contents for sim.