sw: Link rv32 code with only one thread
For quite some time now, on Windows, the bootloader and program builds would randomly hang with what looked to be some kind of deadlock issue in lld. I never spent much time looking into this, but when randomly googling, I found https://github.com/rust-lang/rust/issues/115985#issuecomment-1754112623, which describes similar symptoms and a workaround, which this patch implements.

Admittedly, I haven't spent _that_ much time verifying this, but it appears to work as expected after a few clean/build cycles, so I'll keep going with it for now and revert if the problem comes back again.