rtl: Add additional execute control stages
This brings the number of pipeline stages to 8 instead of 5, which will mean a significant reduction in single-core performance, but should (hopefully!) allow for trivial implementations of mul/div. Additionally, I'm hoping to add fine-grained multithreading, which will bring us up to guaranteed one instruction retired each cycle (assuming full thread utilization), which is a good fit for the kinds of embarassingly-parallel workloads we need to accelerate for the console.